update Android.mk for compat zImage && test dt.img exists or not for compile anykernel.zip

+++ Android.mk
@@ -6,7 +6,8 @@ ak_install_zip_path := $(LOCAL_PATH)/prebuilt_AnyKernel
 
 ANYKERNEL_ZIP_TARGET := $(PRODUCT_OUT)/anykernel
 
-kernel_binary := $(PRODUCT_OUT)/kernel
+ak_kernel_binary := $(PRODUCT_OUT)/kernel
+ak_2nd_kernel_bianry := $(PRODUCT_OUT)/zImage
 dtbimg	      := $(PRODUCT_OUT)/dt.img
 system_module := $(TARGET_OUT)/lib/modules
 ANYKERNEL_INS_DIR := $(PRODUCT_OUT)/anykernel_installer
@@ -28,9 +29,10 @@ $(ANYKERNEL_ZIP_TARGET): bootimage
 	mkdir -p $(ANYKERNEL_INS_DIR)
 
 	cp -a $(ak_install_zip_path)/* $(ANYKERNEL_INS_DIR)/
-	cp -a $(kernel_binary) $(ANYKERNEL_INS_DIR)/kernel
-	cp -a $(dtbimg) $(ANYKERNEL_INS_DIR)/dt.img
-	cp -a $(system_module)/*.ko $(system_module_ins_path)/
+	if [ -f $(ak_kernel_binary) ]; then cp -a $(ak_kernel_binary) $(ANYKERNEL_INS_DIR)/kernel; fi;
+	if [ -f $(ak_2nd_kernel_bianry) ]; then cp -a $(ak_2nk_kernel_binary) $(ANYKERNEL_INS_DIR)/zImage; fi;
+	if [ -f $(dtbimg) ]; then  cp -a $(dtbimg) $(ANYKERNEL_INS_DIR)/dt.img; fi;
+	if [ -d $(system_module) ]; then  cp -rpa $(system_module) $(system_module_ins_path)/../; fi;
 	rm -f $(ANYKERNEL_ZIP_TARGET).zip $(ANYKERNEL_ZIP_TARGET)-unsigned.zip $(TARGET_DEVICE)-$(ANYKERNEL_ZIP_TARGET).zip
 	cd $(ANYKERNEL_INS_DIR) && zip -qr ../$(notdir $@)-unsigned.zip *
 	java -jar $(HOST_OUT_JAVA_LIBRARIES)/signapk.jar $(DEFAULT_SYSTEM_DEV_CERTIFICATE).x509.pem $(DEFAULT_SYSTEM_DEV_CERTIFICATE).pk8 $(ANYKERNEL_ZIP_TARGET)-unsigned.zip $(ANYKERNEL_ZIP_TARGET).zip
